add tests for validation methods related to start vote; move helper functions to testing.go
